Trying to figure out boolean's and relational operators in python
Basically this code is of me trying to figure out how to use boolean.
The top part is just a small warm up of printing numbers divisible by
fives. The lower part is where things get hairy. I'm trying to use user
input as a string to tell if it's raining yes or no. Then the code is
supposed to set goodWeather to true or false and output an appropriate
response, however, that didn't work. It did work when I set user input
equal to an integer, then have the conditional create a goodWeather
boolean variable and evaluate the appropriate response with that. Not
sure yet why it doesn't work with string inputs.